fix(mod6): agregar overlay para pi5

+++ clases/modulo_6/p1_esp32_link/esp32_link.c
@@ -50,8 +50,8 @@ static bool esperando_get;
 
 /* ===================== serdev: recepcion ===================== */
 
-static int esp32_receive_buf(struct serdev_device *serdev,
-                              const unsigned char *data, size_t count)
+static size_t esp32_receive_buf(struct serdev_device *serdev,
+                                 const unsigned char *data, size_t count)
 {
     size_t i;
